Add files via upload
This is supposedly supposed to find links between entities. However, it takes very long to process the file, so I have only managed to test it out with 10 rows and so far it has returned no links between the entities. The slowness might be due to MacOS not having CUDA support for PyTorch, so I can only run the model using my CPU rather than GPUs, so if you are running a WindowsOS, then you can add CUDA support for PyTorch and see if it can run at a reasonable amount of time